pyproject.toml
[project]
name = "dg-bot"
version = "0.1.4"
description = "dg-bot"
readme = "README.md"
requires-python = ">=3.10, <4.0"
dependencies = [
    "nonebot2[fastapi,websockets]>=2.4.4",
    "nonebot-adapter-onebot>=2.4.6",
    "nonebot-plugin-localstore>=0.6.0",
    "httpx>=0.27.0",
]

[tool.nonebot]
plugin_dirs = ["src/plugins"]
builtin_plugins = []

[tool.nonebot.adapters]
nonebot-adapter-onebot = [
    { name = "OneBot V11", module_name = "nonebot.adapters.onebot.v11" }
]
"@local" = []

[tool.nonebot.plugins]
"@local" = []
